// packages/api-client/src/generated/models/githubComSkryfonEmployee360BackendInternalTypesAuthLoginResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Employee360 API
 * REST API for Employee360, an open-source, self-hostable, multi-tenant employee platform. Includes unversioned operational endpoints (health checks) and the versioned client-facing API under /api/v1.
 * OpenAPI spec version: 1.0
 */
import type { GithubComSkryfonEmployee360BackendInternalDomainEntityUser } from './githubComSkryfonEmployee360BackendInternalDomainEntityUser.ts';

export interface GithubComSkryfonEmployee360BackendInternalTypesAuthLoginResponse {
  access_token?: string;
  expires_at?: string;
  refresh_token?: string;
  token_type?: string;
  user?: GithubComSkryfonEmployee360BackendInternalDomainEntityUser;
}