// packages/api-client/src/generated/models/githubComSkryfonEmployee360BackendInternalTypesInvitationInvitationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Employee360 API
 * REST API for Employee360, an open-source, self-hostable, multi-tenant employee platform. Includes unversioned operational endpoints (health checks) and the versioned client-facing API under /api/v1.
 * OpenAPI spec version: 1.0
 */
import type { GithubComSkryfonEmployee360BackendInternalDomainEntityInvitationStatus } from './githubComSkryfonEmployee360BackendInternalDomainEntityInvitationStatus.ts';

export interface GithubComSkryfonEmployee360BackendInternalTypesInvitationInvitationResponse {
  accepted_at?: string;
  created_at?: string;
  department_id?: string;
  email?: string;
  expires_at?: string;
  id?: string;
  invited_by?: string;
  position_id?: string;
  revoked_at?: string;
  role_id?: string;
  status?: GithubComSkryfonEmployee360BackendInternalDomainEntityInvitationStatus;
}